Update EasyDropdown library version to 1.3.2
The current update moves the EasyDropdown library to version 1.3.2. The changes include a fix for an issue where bold font got incorrectly applied to custom text styling, changing the version of the Flutter lints library, and restructuring the logic for the titleWidget style in the easy_dropdown_tile.dart file.